Fixed memory issue with WRRequestDownload under iOS7
Fixed issue which occurs when NSData is copied in WRRequestDownload
under iOS7 which led numerous calls to vm_allocate function. As a
result, application gets low memory warnings and finally gets
terminated. Looks like the true reason of this issue is a bug in iOS 7
SDK, and this workaround fixes it.